round_1b/requirements.txt
PyMuPDF
numpy<2
scikit-learn
spacy
langdetect
networkx
sentence-transformers
pyspellchecker
